+++ cluster/pulumi/common/src/utils.ts
@@ -259,3 +259,10 @@ export const artifactsRepository = config.optionalEnv('SPLICE_ARTIFACTS_REPOSITO
 // One can force splitting them by setting SPLIT_POSTGRES_INSTANCES to true.
 export const SplitPostgresInstances =
   config.envFlag('SPLIT_POSTGRES_INSTANCES') || config.envFlag('ENABLE_CLOUD_SQL');
+
+export const clusterProdLike = config.envFlag('GCP_CLUSTER_PROD_LIKE');
+
+// During development we often overwrite the same tag so we use imagePullPolicy: Always.
+// Outside of development, we use the default which corresponds to IfNotPresent
+// (unless the tag is LATEST which it never is in our setup).
+export const imagePullPolicy = clusterProdLike ? {} : { imagePullPolicy: 'Always' };